Move findById from Factory to Account
